requirements.txt
python-telegram-bot==20.7
asyncpg==0.31.0
python-dotenv==1.0.0
PyNaCl==1.5.0
base58==2.1.1
aiohttp==3.9.5
# WebApp server (optional — only needed if WEBAPP_URL is configured)
fastapi==0.111.0
uvicorn[standard]==0.29.0
